fix stuff

+++ app/src/main/java/dev/keego/musicplayer/remote/lrclib/BestMatchResultPOJO.kt
@@ -5,7 +5,7 @@ data class BestMatchResultPOJO(
     val trackName: String,
     val artistName: String,
     val albumName: String,
-    val duration: Int,
+    val duration: Float,
     val instrument: Boolean,
     val plainLyrics: String,
     val syncedLyrics: String,
